[FARM-277] all Default styles now are private, added a spacing property to RoundedButton style for spacing the icon + text on a button
